+++ config/plugin/visualizer.ts
@@ -2,11 +2,11 @@
  * Generation packaging analysis
  * 生成打包分析
  */
-import visualizer from 'rollup-plugin-visualizer';
 import { isReportMode } from '../utils';
 
-export default function configVisualizerPlugin() {
+export default async function configVisualizerPlugin() {
   if (isReportMode()) {
+    const { visualizer } = await import('rollup-plugin-visualizer');
     return visualizer({
       filename: './node_modules/.cache/visualizer/stats.html',
       open: true,